added api changes
